+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] (2021-04-16)
+
 ### BREAKING CHANGES
 
 - Minimal typescript version is now 4.2 due to the dependency to [typed-query-selector](https://github.com/g-plane/typed-query-selector).
@@ -151,7 +153,8 @@ This is merely a documentation update.
 
 [docs]: https://jjwesterkamp.github.io/lambda-dom/
 
-[Unreleased]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/JJWesterkamp/lambda-dom/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.3.1...v2.0.0
 [1.3.1]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.1.2...v1.2.0